# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Retail Expansion Intelligence System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets:
    get:
      operationId: listMarkets
      tags: [markets]
      summary: List markets
      parameters:
        - name: level
          in: query
          schema: { type: ["string", "null"] }
        - name: parentId
          in: query
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
    post:
      operationId: createMarket
      tags: [markets]
      summary: Add a market
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
  /markets/{id}:
    get:
      operationId: getMarket
      tags: [markets]
      summary: Get market details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
    patch:
      operationId: updateMarket
      tags: [markets]
      summary: Update market
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
    delete:
      operationId: deleteMarket
      tags: [markets]
      summary: Delete market
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /markets/{id}/ranking:
    get:
      operationId: getMarketRanking
      tags: [markets]
      summary: Get ranked locations within a market
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRanking"

  # ─── Competitors ─────────────────────────────────────────────
...
components:
  schemas:
...
    LocationRanking:
      type: object
      required: [id, name, city, state, overallScore, recommendation, riskLevel]
      properties:
        id:           { type: integer }
        name:         { type: string }
        city:         { type: string }
        state:        { type: string }
        overallScore: { type: number }
        demandScore:  { type: ["number", "null"] }
        competitionScore: { type: ["number", "null"] }
        logisticsScore: { type: ["number", "null"] }
        recommendation: { type: string }
        riskLevel:    { type: string }
        confidenceLevel: { type: string }
        rank:         { type: integer }

    # ─── Market ────────────────────────────────────────────────
    Market:
      type: object
      required: [id, name, level, state, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        level:       { type: string, enum: [country, state, district, city, town, ward, micro_market] }
        state:       { type: string }
        district:    { type: ["string", "null"] }
        city:        { type: ["string", "null"] }
        parentId:    { type: ["integer", "null"] }
        population:  { type: ["number", "null"] }
        populationDensity: { type: ["number", "null"] }
        avgIncome:   { type: ["number", "null"] }
        urbanizationRate: { type: ["number", "null"] }
        growthRate:  { type: ["number", "null"] }
        locationCount: { type: integer }
        avgScore:    { type: ["number", "null"] }
        notes:       { type: ["string", "null"] }
        createdAt:   { type: string }
    MarketInput:
      type: object
      required: [name, level, state]
      properties:
        name:         { type: string }
        level:        { type: string }
        state:        { type: string }
        district:     { type: string }
        city:         { type: string }
        parentId:     { type: integer }
        population:   { type: number }
        populationDensity: { type: number }
        avgIncome:    { type: number }
        urbanizationRate: { type: number }
        growthRate:   { type: number }
        notes:        { type: string }
    MarketUpdate:
      type: object
      properties:
        name:         { type: string }
        population:   { type: number }
        populationDensity: { type: number }
        avgIncome:    { type: number }
        urbanizationRate: { type: number }
        growthRate:   { type: number }
        notes:        { type: string }

    # ─── Competitor ────────────────────────────────────────────
...