# lib/api-zod/src/generated/types/locationRanking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retail Expansion Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationRanking {
  id: number;
  name: string;
  city: string;
  state: string;
  overallScore: number;
  /** @nullable */
  demandScore?: number | null;
  /** @nullable */
  competitionScore?: number | null;
  /** @nullable */
  logisticsScore?: number | null;
  recommendation: string;
  riskLevel: string;
  confidenceLevel?: string;
  rank?: number;
}