# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retail Expansion Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketLevel } from './marketLevel';

export interface Market {
  id: number;
  name: string;
  level: MarketLevel;
  state: string;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  population?: number | null;
  /** @nullable */
  populationDensity?: number | null;
  /** @nullable */
  avgIncome?: number | null;
  /** @nullable */
  urbanizationRate?: number | null;
  /** @nullable */
  growthRate?: number | null;
  locationCount?: number;
  /** @nullable */
  avgScore?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}