# lib/api-zod/src/generated/types/marketInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retail Expansion Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketInput {
  name: string;
  level: string;
  state: string;
  district?: string;
  city?: string;
  parentId?: number;
  population?: number;
  populationDensity?: number;
  avgIncome?: number;
  urbanizationRate?: number;
  growthRate?: number;
  notes?: string;
}